Remove link callbacks from main editor, move to Toolbar
